Refactor: broke main page down into smaller components
